+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Changelog
 1.0.0 (unreleased)
 ==================
 
+* Added additional settings
 * Cleaned up file structure
 * Removed Django < 1.8 support
 * Removed ``alt`` attribute and migrated data to Filer